Call url() with suitable options when generating $url from field_link
This follows the code in theme_link_formatter_link_plain() from
link.module. When the raw internal 'url' property is simply
printed, it leads to any query parameters being stripped (and I'm
assuming e.g. that URL fragments are similarly missed off) because
of the internals of how Link sanitizes URLs.

This e.g. makes it impossible to have a carousel item linking to
a YouTube video.